[fix][bsp_common/fatfs] fix pointer assigning type in disk_ioctl function when cmd equal GET_SECTOR_SIZE or GET_BLOCK_SIZE

+++ bsp/bsp_common/fatfs/fatfs_spi_sd.c
@@ -53,12 +53,12 @@ int sd_disk_ioctl(BYTE cmd, void *buff)
             break;
 
         case GET_SECTOR_SIZE:
-            *(DWORD *)buff = SD_CardInfo.CardBlockSize;
+            *(WORD *)buff = SD_CardInfo.CardBlockSize;
             result = RES_OK;
             break;
 
         case GET_BLOCK_SIZE:
-            *(WORD *)buff = SD_CardInfo.CardBlockSize;
+            *(DWORD *)buff = SD_CardInfo.CardBlockSize;
             result = RES_OK;
             break;
